Refactor OpenResponse to use BaseQuestionComponent
[#172821069]
